Calculate Limits of Detection

Calculates limit of detection (LoD) for each target based on the negative controls (blanks). LoD = mean(blanks) + 3*SD(blanks). Designates data as either above or below LoD. Option to specify minimum count threshold for detectability.

lod(
  data_matrix,
  blanks,
  min_count = 0,
  min_blank_no = 4,
  mad_threshold = 2.5,
  ignore_target_blank = NULL,
  targetNoOutlierDetection = NULL,
  match_matrix = NULL
)

Arguments

data_matrix

The Data matrix output from readNULISAseq.R or normalized data from normalization functions.

blanks

Column indices or column names of the blanks in the data_matrix.

min_count

Optional count threshold to apply in addition to the LoD. Default is 0.

min_blank_no

Optional numeric parameter defining the minimum number of blanks required to enable MAD outlier detection and removal. Default is 4.

mad_threshold

Optional numeric parameter defining the threshold used for MAD outlier identification. Default is 2.5.

ignore_target_blank

List of targets and corresponding blanks/NCs to exclude. Names of the list represent targets and values which are arrays represent blank/NC names to ignore during LoD calculation.

targetNoOutlierDetection

Option to provide targets which should NOT have outlier detection applied Defaults to NULL.

match_matrix

Matrix of indices provided by calcSampleTargetNAs. Lists samples/targets that should not be reported

LOD

Vector of limits of detection.

aboveLOD

Logical matrix indicating whether counts are above or below LoD for that target.

Value

A list.