Merge a list of NULISAseq plate data objects

This function merges the outputs from multiple NULISAseq plate runs (as produced by process_loadNULISAseq), combining their metadata, sample, target, QC, and data matrices into a single unified object for downstream analysis.

mergeNULISAseq(dataList, fileNameList, sample_group_covar = "SAMPLE_MATRIX")

Arguments

dataList

A list of data objects, each produced by process_loadNULISAseq, representing individual NULISAseq plates.

fileNameList

A list of filenames corresponding to each plate in dataList.

sample_group_covar

Character, the covariate name used for sample grouping in recalculate multiple runs quantifiability. Default is 'SAMPLE_MATRIX', Function will check first to be sure that the variable is present in the column names of the samples matrix. Can be set to NULL to not use this feature.

Value

A named list containing:

plateID

Character vector of plate IDs

fileNames

Character vector of file names

covariateNames

Character vector of covariate (column) names

ExecutionDetails

List of execution details for each plate

RunSummary

Data frame summarizing each run

IC

Character, the internal control used

targets

Data frame of target information

samples

Data frame of sample information

qcSample

Data frame of sample-level QC metrics

qcTarget

Data frame of target-level QC metrics (if present)

qcPlate

Data frame of plate-level QC metrics

aqParams

Data frame of AQ target parameters (if present)

excluded

Vector of excluded targets (if any)

Data_IC, Data_IClog2, Data_raw, Data_rawlog2, aboveLOD, Data_AQ, Data_AQlog2, Data_Reverse, Data_Reverselog2, Data_AQ_pgmL, Data_AQlog2_pgmL

Matrices of merged data (if present)

unit

Character, AQ unit (if present)

detectability

Data frame of target detectability by sample group/matrix of combined runs.

quantifiability

Data frame of target quantifiability by sample group/matrix of combined runs (if AQ data present).

Details

- Checks for consistent internal control (IC) across plates. - Handles duplicate sample names by appending plate IDs. - Cleans covariate names and merges all available data matrices. - Excludes targets not present in all plates. - Data_IC: IC-normalized data - Data_raw: un-normalized data - Data_AQ: AQ data in aM - Data_AQ_pgmL: AQ data in pgmL - Data_Reverse: Data matrix for reverse curve targets - aboveLOD: Above LOD matrix

Examples

# merged <- mergeNULISAseq(list_of_plate_data, list_of_filenames)