Draw Plate Heatmap

Makes heatmap for a specified target or total counts.

Usage

plateHeatmap(
  target_data,
  well_order = NULL,
  print_counts = TRUE,
  cex = 0.6,
  cex.axis = 0.8,
  digits = 0,
  title = NULL,
  relative = F
)

Arguments

target_data

A numeric vector that represents the counts that will be shown on the heatmap.

well_order

Optional. A numeric vector used to order the wells from A01-A12, B01-B12, ..., H01-H12. If not provided, then it is assumed that the target_data is already in this order.

print_counts

A TRUE / FALSE that specifies whether or not counts will be printed in each cell of the heatmap. Default is TRUE.

cex

Character expansion factor that controls size of the printed counts. Default is 0.7.

cex.axis

Character expansion factor that controls size of the axis labels. Default is 0.8.

digits

number of digits after the decimal point to print. Default is 0.

title

Optional. Plot title.

Value

Draws a heatmap.