Read NULISAseq XML

Reads NULISAseq XML file

readNULISAseq(
  file,
  plateID = NULL,
  file_type = "xml_no_mismatches",
  target_column_names = NULL,
  sample_column_names = NULL,
  sample_group_covar = NULL,
  IC = NULL,
  IPC = NULL,
  SC = NULL,
  NC = NULL,
  Bridge = NULL,
  Calibrator = NULL,
  replaceNA = TRUE
)

Arguments

file

Character string. Path and name of the file.

plateID

Character string that denotes plate ID.

file_type

Character string. Type of input file. Options include xml_no_mismatches, xlsx (output from ACC) or csv_long (output from writeNULISAseq() function).

target_column_names

A vector of column names for target-specific data that will appear in the target data frame output. Only used for csv_long format, and only needed when target column names are different from the default, which are c('Target', 'AlamarTargetID', 'UniProtID', 'ProteinName').

sample_column_names

A vector of column names for sample-specific data that will appear in the sample data frame output. Only used for csv_long format, and only needed when sample column names are different from the default. Default includes all column names in the csv except for the target column names, given above (either the default or whatever is specified), and the following: c('Panel', 'PanelLotNumber', 'LOD', 'NPQ'). The sample-specific data will include by default c('PlateID', 'SampleName', 'SampleType', 'SampleQC'). In addition, by default, it will include any other variables in the dataset. If sample_column_names is specified, however, it will only include those variables in the sample data frame output.

sample_group_covar

string that represents the name of the covariate that defines sample group information. Defaults to NULL. Typically, this covariate defines whether each sample is plasma, serum, csf, etc. This information can be used when reporting detectability. For instance report sample group specific detectability values, etc.).

IC

string(s) that represents the names of internal control targets. Default is 'mCherry'.Only used for xml file formats.

IPC

string(s) present in the sample names that represent the inter-plate control wells. For example, 'IPC'. Set to NULL (default) to use the type variable from the Barcode B file or if there are no IPCs. Only used for xml file formats.

SC

string(s) present in the sample names that represent the sample control wells. For example, 'SC'. Set to NULL (default) to use the type variable from the Barcode B file or if there are no IPCs. Only used for xml file formats.

NC

string(s) present in the sample names that represent the negative control wells. For example, 'NC'. Set to NULL (default) to use the type variable from the Barcode B file or if there are no NCs. Only used for xml file formats.

Bridge

string(s) present in the sample names that represent the bridge sample wells. Set to NULL (default) to use the type variable from the Barcode B file or if there are no bridge samples (default). Only used for xml file formats.

Calibrator

string(s) present in the sample names that represent the calibrator wells. Set to NULL (default) to use the type variable from the Barcode B file or if there are no calibrator samples (default). Only used for xml file formats.

replaceNA

Logical. If TRUE (default), will replace missing counts with zeros.

Value

List of lists, data frames, and matrices. Output will differ slightly depending on the input file type.