Write Plate QC table

Writes NULISAseq QC plate flags

QCFlagPlate(
  raw,
  normed,
  aboveLOD,
  targets,
  samples,
  ICs = NULL,
  IPCs = NULL,
  NCs = NULL,
  SCs = NULL,
  AQ = TRUE,
  AQ_QC = NULL,
  Sample_QC = NULL
)

Arguments

raw

unnormalized data

normed

normalized data

aboveLOD

aboveLOD from lod function

ICs

names of IC samples

IPCs

names of IPC samples

NCs

names of NC samples

SCs

names of SC samples

AQ

Whether this assay is AQ

AQ_QC

AQ QC metrics, these are necessary to calculate a plate QC metric for AQ

Sample_QC

Sample QC metrics, one of these is necessary to calculate a plate QC metric for AQ

Value

QC table

Examples

# QCFlagPlate(inputtable)